requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
amqp==5.3.1
appdirs==1.4.4
asgiref==3.8.1
async-timeout==5.0.1
asyncio==3.4.3
attrs==24.2.0
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
dj-database-url==2.3.0
Django==5.1.3
django-celery-beat==2.7.0
django-environ==0.11.2
django-timezone-field==7.0
frozenlist==1.5.0
gunicorn==23.0.0
h11==0.14.0
idna==3.10
kombu==5.4.2
multidict==6.1.0
packaging==24.2
prompt_toolkit==3.0.48
propcache==0.2.1
psycopg2-binary==2.9.10
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.2
sync-to-async==0.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
yarl==1.18.3
